[Windows] Fix nntrainer double to float init warnings
This PR fix some of compilation warnings related with initialization of float variable with double value on Windows build

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/models/dynamic_training_optimization.h
@@ -57,7 +57,7 @@ public:
   /**
    * @brief     Constructor of DynamicFineTuning Optimization
    */
-  DynamicTrainingOptimization(int threshold_ = 1, int skip_n_iter = 1);
+  DynamicTrainingOptimization(float threshold_ = 1.0f, int skip_n_iter = 1);
 
   /**
    * @brief     Set threshold for optimization
